requirements.txt
discord.py==2.3.1
Pillow==10.0.0
openai==0.27.0
requests==2.31.0